+++ include/structSnakeNode.h
@@ -8,10 +8,10 @@
 
 typedef struct SnakeNode
 {
-    int node_type;      // 0 = tail, 1 = head, 2 = bend
-    int x;              // x coord of entity
-    int y;              // y coord of entity
-    int direction;      // 1 - Up, 2 - Left, 3 - Down, 4 - Right
+    uint8_t node_type;      // 0 = tail, 1 = head, 2 = bend
+    uint8_t x;              // x coord of entity
+    uint8_t y;              // y coord of entity
+    uint8_t direction;      // 1 - Up, 2 - Left, 3 - Down, 4 - Right
     SnakeNode *next;    // Pointer to the next node in the list
 
 } SnakeNode;
